fix: cast to int64 before MaxUint32 comparison for 32-bit ARM
On 32-bit platforms, int is 32 bits and math.MaxUint32 (4294967295)
overflows int, causing a compile error. Cast to int64 first so the
comparison works on all architectures.

+++ pkg/loghttp/params.go
@@ -32,7 +32,7 @@ func limit(r *http.Request) (uint32, error) {
 	if l <= 0 {
 		return 0, errors.New("limit must be a positive value")
 	}
-	if l > math.MaxUint32 {
+	if int64(l) > math.MaxUint32 {
 		return 0, fmt.Errorf("limit value %d is out of range [0, %d]", l, uint64(math.MaxUint32))
 	}
 	return uint32(l), nil
@@ -46,7 +46,7 @@ func lineLimit(r *http.Request) (uint32, error) {
 	if l <= 0 {
 		return 0, errors.New("limit must be a positive value")
 	}
-	if l > math.MaxUint32 {
+	if int64(l) > math.MaxUint32 {
 		return 0, fmt.Errorf("line_limit value %d is out of range [0, %d]", l, uint64(math.MaxUint32))
 	}
 	return uint32(l), nil
@@ -67,7 +67,7 @@ func detectedFieldsLimit(r *http.Request) (uint32, error) {
 	if l <= 0 {
 		return 0, errors.New("limit must be a positive value")
 	}
-	if l > math.MaxUint32 {
+	if int64(l) > math.MaxUint32 {
 		return 0, fmt.Errorf("limit value %d is out of range [0, %d]", l, uint64(math.MaxUint32))
 	}
 	return uint32(l), nil
@@ -155,7 +155,7 @@ func tailDelay(r *http.Request) (uint32, error) {
 	if err != nil {
 		return 0, err
 	}
-	if l < 0 || l > math.MaxUint32 {
+	if l < 0 || int64(l) > math.MaxUint32 {
 		return 0, fmt.Errorf("delay_for value %d is out of range [0, %d]", l, uint64(math.MaxUint32))
 	}
 	return uint32(l), nil
